requirements.txt
# Core Framework
streamlit==1.29.0
uagents==0.22.5

# AI & NLP
groq
openai

# Voice Processing
SpeechRecognition
pyttsx3
streamlit-mic-recorder

# HTTP Requests
requests

# Data Processing
pandas
numpy

# Database
neo4j-driver

# Visualization
plotly

# Environment
python-dotenv

# Audio Processing
pyaudio
sounddevice

# Date/Time
python-dateutil

# Additional Utilities
json5
typing-extensions
